requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
idna==3.6
pillow==10.1.0
PyJWT==2.8.0
pypdf==3.17.4
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
gunicorn
pymupdf

# Interview Agent Dependencies
langgraph>=0.6.0,<0.7.0
langchain-openai>=0.2.1
langchain-anthropic>=0.2.1
langchain-groq>=0.1.0
langchain>=0.3.8
langchain-core>=0.3.8
langgraph-sdk>=0.1.32
sentence-transformers>=2.2.0
torch>=2.0.0
langgraph-checkpoint-sqlite
